Feature: Delete product endpoint and test

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/bancolombia/r2dbc/entities/FranchiseEntity.java
@@ -5,17 +5,17 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
-    @Data
-    @ToString
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Table("franchise")
+@Data
+@ToString
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Table("franchise")
 public class FranchiseEntity {
 
     @Id
     @Column("franchise_id")
-    private Long franchiseId;
+    private Integer franchiseId;
     @Column("name")
     private String name;
 }
